% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-query.R
\name{create_search_url}
\alias{create_search_url}
\title{Create URL}
\usage{
create_search_url(search_url, parameters = NULL)
}
\arguments{
\item{search_url}{A URL used to construct the API query}

\item{parameters}{A named list specifying API query parameters}
}
\value{
A URL containing a query string
}
\description{
Appends a query string to a URL
}
\keyword{internal}
