% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{DREAM}
\alias{DREAM}
\title{FIMM-UTU DREAM winning implementation of an ensemble of Penalized Cox Regression models for mCPRC research (ePCR)}
\format{An object of class \code{PEP} of length 1.}
\usage{
data(ePCRmodels)
}
\description{
FIMM-UTU DREAM winning implementation of an ensemble of Penalized Cox Regression models for mCPRC research (ePCR)
}
\note{
Notice that in order to save space, some slots in the S4 object have been set to null.
}
\references{
Guinney J, Wang T, Laajala TD, et al. Prediction of overall survival for patients with metastatic castration-resistant prostate cancer: development of a prognostic model through a crowdsourced challenge with open clinical trial data. Lancet Oncol 2017; 18: 132-142.
}
\author{
Teemu Daniel Laajala \email{teelaa@utu.fi}
}
\keyword{datasets}
