% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-query.R
\name{get_data_for_lmk_key}
\alias{get_data_for_lmk_key}
\alias{get_epc_certificate}
\alias{get_epc_recommendations}
\title{Retrieves data for specific LMK key}
\usage{
get_data_for_lmk_key(record_type, api_type, lmk_key)

get_epc_certificate(record_type, lmk_key)

get_epc_recommendations(record_type, lmk_key)
}
\arguments{
\item{record_type}{One of "domestic", "non-domestic" or "display"}

\item{api_type}{One of "certificate" or "recommendations"}

\item{lmk_key}{An LMK key corresponding to a specific EPC certifcate}
}
\value{
A list object of class 'epc_api' holding a data frame containing the data and
a list object containing the API response data.
}
\description{
Queries EPC API to retrieve either recommendations or a certificate for a specific
LMK key. \code{get_epc_certificate} and \code{get_epc_recommendations} are wrapper
functions for \code{get_data_for_lmk_key} with \code{api_type} set to "certificate"
and "recommendations" respectively.
}
\details{
Note that this function will prompt you for a valid API key. You will need a registered account before you can
access the data. See the \href{https://epc.opendatacommunities.org/docs/api}{documentation} for further guidance.
}
\examples{
\dontrun{
get_data_for_lmk_key("domestic", "certificate", "887164559962017031316220145408873")
}

\dontrun{
get_epc_certificate("domestic", "887164559962017031316220145408873")
}

\dontrun{
get_epc_recommendations("domestic", "887164559962017031316220145408873")
}

}
