% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{meanrank}
\alias{meanrank}
\title{Compute mean of predicted risk ranks for an ePCR ensemble}
\usage{
meanrank(x)
}
\arguments{
\item{x}{A list or a matrix of risk scores given per each ensemble member (each column or list member is considered an equal member of the ensemble)}
}
\value{
An averaged predicted risk rank over all the ensemble members
}
\description{
Compute mean of predicted risk ranks for an ePCR ensemble
}
\note{
Extensively called by the 'predict'-function for PEP-objects when risk predictions are performed over the ensemble
}
\author{
Teemu Daniel Laajala \email{teelaa@utu.fi}
}
