% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{normriskrank}
\alias{normriskrank}
\title{Normalize ensemble risk scores to ranks and then to uniform range}
\usage{
normriskrank(x)
}
\arguments{
\item{x}{A list or a matrix of risk scores given per each ensemble member (each column or list member is considered an equal member of the ensemble}
}
\value{
An averaged predicted risk rank over all the ensemble members that has been normalized to the range [0,1] based on: (x - min(x)) / (max(x) - min(x)) -> [0,1]
}
\description{
Normalize ensemble risk scores to ranks and then to uniform range
}
\note{
Normalizes 'predict'-function calls for PEP-objects after calling 'meanrank'-function
}
\author{
Teemu Daniel Laajala \email{teelaa@utu.fi}
}
