% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep.R
\docType{methods}
\name{PEP-methods}
\alias{PEP-methods}
\alias{print,PEP-method}
\alias{predict,PEP-method}
\title{PEP-methods}
\usage{
\S4method{print}{PEP}(x, ...)

\S4method{predict}{PEP}(object, type = "response", newx)
}
\arguments{
\item{x}{Generic x}

\item{...}{Additional custom parameters passed on}

\item{object}{PEP-ensemble model object}

\item{type}{Type of prediction; either "response" or "ensemble"}

\item{newx}{New data matrix}
}
\description{
PEP-methods

Print a PEP object

predict.PEP: Predict for a novel patient from current PEP-ensemble
}
