% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-results.R
\name{process_response}
\alias{process_response}
\title{Parses EPC API response}
\usage{
process_response(parsed_resp)
}
\arguments{
\item{resp}{A response object returned by \code{httr::GET}}
}
\value{
A list of class 'epc_api' containing a data frame holding the results of the
query ("contents") and a response object ("response")
}
\description{
Parses EPC API response
}
\keyword{internal}
