% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-query.R
\name{query_epc_data}
\alias{query_epc_data}
\title{Query EPC web API}
\usage{
query_epc_data(url)
}
\arguments{
\item{url}{URL for the EPC API}
}
\value{
A list containing a list object of content data, and
a list object containing the API response data.
}
\description{
Makes an HTTP get request to retrieve EPC data from a user specified URL.
}
\examples{
\dontrun{
# Default domestic query
query_epc_data("https://epc.opendatacommunities.org/api/v1/domestic/search")
}

}
