% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-query.R
\name{run_epc_query}
\alias{run_epc_query}
\title{Run EPC query}
\usage{
run_epc_query(url)
}
\arguments{
\item{url}{URL for the EPC API}
}
\value{
A list of class 'epc_api' containing a data frame holding the results of the
query ("contents") and a response object ("response")
}
\description{
Run EPC query
}
\keyword{internal}
