% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-query.R
\name{search_epc_data}
\alias{search_epc_data}
\title{Search EPC data}
\usage{
search_epc_data(record_type, ..., size = NULL, paginated = FALSE)
}
\arguments{
\item{record_type}{One of "domestic", "non-domestic" or "display"}

\item{...}{Named query parameters}

\item{size}{A query parameter indicating the size of the desired result set. If \code{paginated = TRUE},
this indicates the desired number of results per page. The API defaults to a page size of 25 if
not specified.}

\item{paginated}{A Boolean value indicating whether a paginated result is required.
Note that the API can handle result sets of up to 10,000 with a maximum page size of
5,000. If fewer than 5,000 results are expected, it may be easier to leave this set to
FALSE (the default) and specify a sufficiently large \code{size} parameter to return a non-paginated
result set. Otherwise, set this to TRUE and set the \code{size} argument to to indicate the
number of results per page (up to a maximum of 5,000, which would return two pages for a
query returning more than 5,000 results).}
}
\value{
A list of class 'epc_api' holding a data frame containing the data and
a list containing API response objects for each page.
}
\description{
Queries the EPC API to retrieve EPC data filtered for user-specified
parameters.
}
\examples{
search_epc_data("domestic", postcode = "SW1A")
search_epc_data("domestic", postcode = "TW1 4PG", size = 2, paginated = TRUE)

}
