% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-query.R
\name{validate_query}
\alias{validate_query}
\title{Validate query parameters}
\usage{
validate_query(parameters)
}
\arguments{
\item{parameters}{A named list of query parameters corresponding to filters described
in the EPC API documentation.}
}
\description{
Checks that valid filters are applied.
}
\keyword{internal}
