% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{check_api_key}
\alias{check_api_key}
\title{Function to check API key}
\usage{
check_api_key(reset = FALSE)
}
\arguments{
\item{reset}{A Boolean value indicating whether or not you want to reset the
API key environment variable.}
}
\description{
Checks for an environment variable called 'EPC_API_KEY'. If this is set, it returns to
the parent frame. If it is not set, it prompts the user to set this before continuing.
The function can be called interactively to reset the key.
}
\details{
Note that you should use the Base64-encoded account identifier composed of your email address and api-key as
provided in the \href{https://epc.opendatacommunities.org/docs/api/domestic#using_this_api}{documentation}.
}
