% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-query.R
\name{query_epc_data}
\alias{query_epc_data}
\title{Query EPC web API}
\usage{
query_epc_data(url)
}
\arguments{
\item{url}{URL for the EPC API including the required query string}
}
\value{
A list containing the data retrieved and API response
}
\description{
Makes an HTTP get request to retrieve EPC data from a specified URL.
}
\examples{
\dontrun{
# Default domestic query
query_epc_data("https://epc.opendatacommunities.org/api/v1/domestic/search")
}

}
