% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-query.R
\name{query_epc_data}
\alias{query_epc_data}
\title{Query EPC web API}
\usage{
query_epc_data(url)
}
\arguments{
\item{url}{URL for the EPC API}
}
\value{
A list object of class 'epc_api' holding a data frame containing the data and
a list object containing the API response data.
}
\description{
Makes an HTTP get request to retrieve EPC data.
}
\examples{
query_epc_data("https://epc.opendatacommunities.org/api/v1/domestic/search")

}
